Addition of interface code
Master branch as of July 2024 was modified to enable the interface extraction mode that was first created in the interface branch in December 2023. This original branch was not able to be successfully merged with master, so this is a fresh start. Additional change from the original branch is the new parameter exposing the distance threshold for defining the interface.